fix: clamp autocert_renew_before + 64-bit sched rearm math (#115)
* fix: clamp autocert_renew_before + 64-bit sched rearm math

Audit 2026-07-08 (LOW RISK) surfaced two driver renew-window footguns:

- renew_before was ngx_conf_set_sec_slot with no upper bound. A value at or
  above a cert's lifetime makes "now >= notAfter - renew_before" always true,
  so every sweep reissues -- hammering the CA and tripping ACME rate limits
  (operator self-DoS). Clamp at init_main_conf to >0 && <=89d (LE certs live
  90d), rejecting with NGX_LOG_EMERG at config time, mirroring the existing
  dns_hook_timeout gate.
- sched rearm computed (ngx_msec_t)(soonest-now)*1000, casting before the
  multiply. Safe today (soonest is CAP-bounded to now+1h) but wraps if the CAP
  grows; do the *1000 in 64-bit first to match the renew_before math above.

Extend config-rejects.sh: reject 90d/999999999, accept 30d. All 16
config-reject checks green locally against the rebuilt module.

* test: fix renewal.sh for renew_before 89d clamp

The renew_before clamp (module.c) now rejects values > 89d, but renewal.sh
forced its "always inside renew window" reissue with `autocert_renew_before
9999d` — which the new config gate rejects, failing `nginx -t`.

Instead shorten the Pebble cert lifetime to 1h via a `profiles.default.
validityPeriod: 3600` block (the deprecated top-level certificateValidityPeriod
is ignored by current Pebble, which falls back to the 90d CA default), and set
renew_before to 2h. 2h > 1h lifetime keeps every fresh cert permanently inside
its renew window while staying under the 89d clamp. The M9 negative-path
control (small renew_before => not due) uses 60s, still valid against 1h certs.

* ci: re-trigger (angie 1.12.0 transient upstream, now yanked from mirror)

* fix: build against angie 1.12.0 (resolver connections array->pointer)

angie 1.12.0 reworked ngx_resolver_t: `connections` changed from an
ngx_array_t (mainline nginx, angie < 1.12.0) to a linked list of
ngx_resolver_connection_t* (NULL when no name servers are configured).
The http{}-resolver fallback in copy_srv_conf() read
`clcf->resolver->connections.nelts`, which no longer compiles under -Werror
on angie 1.12.0 (`connections is a pointer`).

Guard the "has a name server?" test on the numeric `angie_version` macro
(from src/core/angie.h, transitively included via ngx_module.h): >= 1012000
uses `connections != NULL`, otherwise the ngx_array_t `.nelts > 0`. Verified
both branches compile clean: nginx 1.31.2 (array) and angie 1.12.0 (pointer).

+++ src/ngx_autocert_driver.c
@@ -972,7 +972,11 @@ rearm:
             }
         }
         if (soonest != 0) {
-            ngx_msec_t  until = (ngx_msec_t) (soonest - now) * 1000;
+            /* 64-bit multiply BEFORE the ngx_msec_t narrow (mirrors the
+             * renew_before math above): soonest is CAP-bounded today so this
+             * cannot wrap, but doing it in 64-bit is regression-proof if the
+             * eligibility CAP ever grows. */
+            ngx_msec_t  until = (ngx_msec_t) ((uint64_t) (soonest - now) * 1000);
             if (until < interval) {
                 interval = until;
             }
